new: provide a direct executable for using gandi.cli without installation
Provides in the root of the repository, a binary to allow to
use gandi.cli without requiring an installation (and copying files
into your site-packages). This streamlines the process for people
wanting to use the developpement version that is directly inside their
GIT repository clone.

Corrected also the shebang of ``gandi/cli/__main__.py`` that was
pointing to the system python, which is not a good idea especially for
people using virtual environments, or any other reason to switch between
multiples version of python on your system.